implement plan editor web interface
